modified:   fsharp/basic_data_structures/basic_data_structures.fsproj 	new file:   fsharp/basic_data_structures/src/History/Figure_history.fs 	renamed:    fsharp/basic_data_structures/src/History/History.fs -> fsharp/basic_data_structures/src/History/Full_history.fs 	modified:   fsharp/basic_data_structures/src/Interval.fs 	modified:   fsharp/basic_data_structures/src/identifiers_in_database.fs 	new file:   fsharp/io/src/visualised/painted/Full_history.fs 	modified:   fsharp/io/src/visualised/painted/History.fs

+++ fsharp/basic_data_structures/src/identifiers_in_database.fs
@@ -2,4 +2,4 @@
 
 type Figure_id = string
 type Vertex_id = string
-type Moment = uint64
+type Moment = int
